+++ pages/user/settings/preferences.vue
@@ -1,7 +1,7 @@
 <template>
     <div class="surface-ground px-4 md:px-6 lg:px-8">
         <div class="p-fluid flex flex-column lg:flex-row">
-            <UserSettingsNav />
+            <PanelNav :links="links" :buttons="buttons" />
             <div v-if="currentUser.props" class="flex-auto">
                 <div class="surface-card p-5 border-1 surface-border border-round">
                     <div class="text-900 font-semibold text-lg">Preferences</div>
@@ -29,4 +29,6 @@ definePageMeta({
     middleware: ['auth']
 });
 const currentUser = useCurrentUserStore();
+const links = userSettingsLinks();
+const buttons = userSettingsButtons();
 </script>