+++ docs/change-log.md
@@ -18,6 +18,7 @@ outline: deep
 * Added ManageButton component.
 * Added link to user in user admin panel.
 * Removed all references to PrimeVue icons as they don't size well with the Iconify icons.
+* Replaced UserSettingsNav with PanelNav and userSettingsLinks() and userSettingsButtons().
 
 ## 2024-04-25
 * Fixed a bug where ving record fields of type 'int' were not being initialized properly.